Fix integration tests after merging development

+++ integrationTests/vm/iele/vmAgar_test.go
@@ -199,7 +199,7 @@ func TestAgarioContractJoinGameReward(t *testing.T) {
 	senderAddressBytes := []byte("12345678901234567890123456789012")
 	senderNonce := uint64(11)
 	senderBalance := big.NewInt(100000000)
-	round := uint32(444)
+	round := uint64(444)
 	gasPrice := uint64(0)
 	gasLimit := uint64(1000000)
 
@@ -327,7 +327,7 @@ func BenchmarkAgarioJoinGame(b *testing.B) {
 	senderAddressBytes := []byte("12345678901234567890123456789012")
 	senderNonce := uint64(11)
 	senderBalance := big.NewInt(100000000)
-	round := uint32(444)
+	round := uint64(444)
 	gasPrice := uint64(0)
 	gasLimit := uint64(1000000)
